+++ jscomp/gentype_tests/typescript-react-example/src/nested/Tuples.gen.tsx
@@ -11,7 +11,7 @@ import * as TuplesBS__Es6Import from './Tuples.bs';
 const TuplesBS: any = TuplesBS__Es6Import;
 
 // tslint:disable-next-line:interface-over-type-literal
-export type coord = [number, number, (null | undefined | number)];
+export type coord = [number, number, (undefined | number)];
 
 // tslint:disable-next-line:interface-over-type-literal
 export type coord2 = [number, number, (null | undefined | number)];
@@ -24,21 +24,15 @@ export type couple = [person, person];
 
 export const testTuple: (param:[number, number]) => number = TuplesBS.testTuple;
 
-export const origin: [number, number, (null | undefined | number)] = TuplesBS.origin;
+export const origin: [number, number, (undefined | number)] = TuplesBS.origin;
 
-export const computeArea: (param:[number, number, (null | undefined | number)]) => number = function (Arg1: any) {
-  const result = TuplesBS.computeArea([Arg1[0], Arg1[1], (Arg1[2] == null ? undefined : Arg1[2])]);
-  return result
-};
+export const computeArea: (param:[number, number, (undefined | number)]) => number = TuplesBS.computeArea;
 
-export const computeAreaWithIdent: (param:coord) => number = function (Arg1: any) {
-  const result = TuplesBS.computeAreaWithIdent([Arg1[0], Arg1[1], (Arg1[2] == null ? undefined : Arg1[2])]);
-  return result
-};
+export const computeAreaWithIdent: (param:coord) => number = TuplesBS.computeAreaWithIdent;
 
 export const computeAreaNoConverters: (param:[number, number]) => number = TuplesBS.computeAreaNoConverters;
 
-export const coord2d: <T1,T2,T3>(x:T1, y:T2) => [T1, T2, (null | undefined | T3)] = function <T1,T2,T3>(Arg1: any, Arg2: any) {
+export const coord2d: <T1,T2,T3>(x:T1, y:T2) => [T1, T2, (undefined | T3)] = function <T1,T2,T3>(Arg1: any, Arg2: any) {
   const result = Curry._2(TuplesBS.coord2d, Arg1, Arg2);
   return result
 };